remove symbols under the haystack.document_stores namespace
