+++ haystack/components/caching/cache_checker.py
@@ -5,7 +5,7 @@ import importlib
 import logging
 
 from haystack import component, Document, default_from_dict, default_to_dict, DeserializationError
-from haystack.document_stores import DocumentStore
+from haystack.document_stores.types import DocumentStore
 
 
 logger = logging.getLogger(__name__)
