+++ haystack/components/writers/document_writer.py
@@ -4,7 +4,7 @@ import importlib
 import logging
 
 from haystack import component, Document, default_from_dict, default_to_dict, DeserializationError
-from haystack.document_stores import DocumentStore, DuplicatePolicy
+from haystack.document_stores.types import DocumentStore, DuplicatePolicy
 
 logger = logging.getLogger(__name__)
