--- haystack/document_stores/__init__.py
@@ -1,12 +1,0 @@
-from haystack.document_stores.protocol import DocumentStore, DuplicatePolicy
-from haystack.document_stores.in_memory.document_store import InMemoryDocumentStore
-from haystack.document_stores.errors import DocumentStoreError, DuplicateDocumentError, MissingDocumentError
-
-__all__ = [
-    "DocumentStore",
-    "DuplicatePolicy",
-    "InMemoryDocumentStore",
-    "DocumentStoreError",
-    "DuplicateDocumentError",
-    "MissingDocumentError",
-]
