+++ haystack/document_stores/errors/__init__.py
@@ -1,0 +1,3 @@
+from .errors import DocumentStoreError, DuplicateDocumentError, MissingDocumentError
+
+__all__ = ["DocumentStoreError", "DuplicateDocumentError", "MissingDocumentError"]
