+++ haystack/document_stores/errors/errors.py
