+++ haystack/document_stores/types/__init__.py
@@ -1,0 +1,4 @@
+from .protocol import DocumentStore
+from .policy import DuplicatePolicy
+
+__all__ = ["DocumentStore", "DuplicatePolicy"]
