+++ haystack/document_stores/types/policy.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class DuplicatePolicy(Enum):
+    NONE = "none"
+    SKIP = "skip"
+    OVERWRITE = "overwrite"
+    FAIL = "fail"
