+++ haystack/pipeline_utils/rag.py
@@ -11,7 +11,8 @@ from haystack.components.embedders import SentenceTransformersTextEmbedder
 from haystack.components.generators import OpenAIGenerator, HuggingFaceTGIGenerator
 from haystack.components.retrievers import InMemoryEmbeddingRetriever
 from haystack.dataclasses import Answer
-from haystack.document_stores import InMemoryDocumentStore, DocumentStore
+from haystack.document_stores.types import DocumentStore
+from haystack.document_stores.in_memory import InMemoryDocumentStore
 
 
 def build_rag_pipeline(
