+++ haystack/testing/document_store.py
@@ -7,7 +7,7 @@ import pytest
 import pandas as pd
 
 from haystack.dataclasses import Document
-from haystack.document_stores import DocumentStore, DuplicatePolicy
+from haystack.document_stores.types import DocumentStore, DuplicatePolicy
 from haystack.document_stores.errors import DuplicateDocumentError
 from haystack.errors import FilterError
