+++ test/components/retrievers/test_in_memory_embedding_retriever.py
@@ -7,7 +7,7 @@ from haystack import Pipeline, DeserializationError
 from haystack.testing.factory import document_store_class
 from haystack.components.retrievers.in_memory_embedding_retriever import InMemoryEmbeddingRetriever
 from haystack.dataclasses import Document
-from haystack.document_stores import InMemoryDocumentStore
+from haystack.document_stores.in_memory import InMemoryDocumentStore
 
 
 class TestMemoryEmbeddingRetriever:
